Fix size_t to int conversion warnings

+++ src/private/httpreader.hxx
@@ -46,7 +46,7 @@ public:
         if ( icollectCapacity > 0    &&    newLength > icollectCapacity )
             return false; // the capacity is full
 
-        ibody.append(data, length);
+        ibody.append(data, static_cast<int>(length));
         return true;
     }
